+++ band-session-extension/config.js
@@ -3,9 +3,9 @@
  */
 
 // 개발용
-// const SERVER_URL = 'http://localhost:3001';
-// const APP_DOMAIN = 'localhost';
+const SERVER_URL = 'http://localhost:3001';
+const APP_DOMAIN = 'localhost';
 
 // 프로덕션용
-const SERVER_URL = 'https://snsauto.abcpharm.net';
-const APP_DOMAIN = 'snsauto.abcpharm.net';
+//const SERVER_URL = 'https://snsauto.abcpharm.net';
+//const APP_DOMAIN = 'snsauto.abcpharm.net';